CGS/requirements.txt
tensorboard
tensorboardX
tqdm
scikit-image
opencv-python
h5py
matplotlib
efficientnet_pytorch
einops
timm
yacs
batchgenerators
medpy